requirements.txt
pytest~=8.3.3
selene~=2.0.0rc9
selenium~=4.29.0
allure-pytest
allure-python-commons~=2.13.5
python-dotenv
webdriver_manager
requests
jsonschema
Appium-Python-Client